Fix for Cantabile loading saved Stochas state
It looks like Cantabile fires up the plugin editor, then destroys it
before loading the saved project. There is a notifier callback that the
plugin calls that is owned by the editor. since it was destroyed the
callback object became invalid. solution is to clear the notifier when
the plugin editor is destroyed so that it doesn't call it.